feat: replace EventCardInputButton with EventCardInput for editable event fields
